requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
Pillow==7.2.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
